Pull out some reusable pieces of pictures
